requirements.txt
absl-py==1.4.0
aiofiles==22.1.0
aiohttp==3.8.4
aiosignal==1.3.1
aiosqlite==0.18.0
antlr4-python3-runtime==4.9.3
anyio==3.6.2
apex @ file:///home/gpu/Kaggle/GOOGLE-ASL/apex
apturl==0.5.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.2.0
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.11.2
bleach==6.0.0
blessed==1.20.0
blinker==1.4
Brlapi==0.8.3
cachetools==5.3.0
certifi==2020.6.20
cffi==1.15.1
chardet==4.0.0
charset-normalizer==3.1.0
click==8.1.3
cloudpickle==2.2.1
cmake==3.25.0
colorama==0.4.4
comm==0.1.2
command-not-found==0.3
contourpy==1.0.7
croniter==1.3.8
cryptography==3.4.8
cupshelpers==1.0
cycler==0.11.0
dask==2023.3.1
dateutils==0.6.12
dbus-python==1.2.18
debugpy==1.6.6
decorator==5.1.1
deepdiff==6.3.0
defer==1.0.6
defusedxml==0.7.1
distro==1.7.0
distro-info===1.1build1
dnspython==2.3.0
email-validator==1.3.1
executing==1.2.0
fastapi==0.88.0
fastjsonschema==2.16.3
filelock==3.9.0
flatbuffers==23.3.3
fonttools==4.39.2
fqdn==1.5.1
frozenlist==1.3.3
fsspec==2023.3.0
google-auth==2.16.3
google-auth-oauthlib==0.4.6
# Editable Git install with no remote (Google-ISLR-Kaggle-competition==1.0)
-e /home/gpu/Kaggle/GOOGLE-ASL
grpcio==1.51.3
h11==0.14.0
httpcore==0.16.3
httplib2==0.20.2
httptools==0.5.0
httpx==0.23.3
idna==3.3
importlib-metadata==4.6.4
inquirer==3.1.3
ipykernel==6.22.0
ipympl==0.9.3
ipython==8.11.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.18.2
jeepney==0.7.1
Jinja2==3.1.2
joblib==1.2.0
json5==0.9.11
jsonpointer==2.3
jsonschema==4.17.3
jupyter-events==0.6.3
jupyter-ydoc==0.2.3
jupyter_client==8.1.0
jupyter_core==5.3.0
jupyter_server==2.4.0
jupyter_server_fileid==0.8.0
jupyter_server_terminals==0.4.4
jupyter_server_ydoc==0.6.1
jupyterlab==3.6.1
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.5
jupyterlab_server==2.20.0
kaggle==1.5.13
keyring==23.5.0
kiwisolver==1.4.4
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lightning==2.0.0
lightning-cloud==0.5.32
lightning-utilities==0.8.0
lit==15.0.7
locket==1.0.0
louis==3.20.0
macaroonbakery==1.3.1
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mediapipe==0.9.1.0
mistune==2.0.5
more-itertools==8.10.0
mpmath==1.2.1
multidict==6.0.4
nbclassic==0.5.3
nbclient==0.7.2
nbconvert==7.2.9
nbformat==5.7.3
nest-asyncio==1.5.6
netifaces==0.11.0
networkx==3.0rc1
notebook==6.5.3
notebook_shim==0.2.2
numpy==1.24.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.0
olefile==0.46
opencv-contrib-python==4.7.0.72
ordered-set==4.1.0
orjson==3.8.8
packaging==23.0
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.3
partd==1.3.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==3.1.1
prometheus-client==0.16.0
prompt-toolkit==3.0.38
protobuf==3.20.3
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==11.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.20.1
pycparser==2.21
pycups==2.0.1
pydantic==1.10.6
Pygments==2.14.0
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyOpenSSL==21.0.0
pyparsing==3.0.9
pyRFC3339==1.1
pyrsistent==0.19.3
python-apt==2.4.0+ubuntu1
python-dateutil==2.8.2
python-debian===0.1.43ubuntu1
python-dotenv==1.0.0
python-editor==1.0.4
python-json-logger==2.0.7
python-multipart==0.0.6
python-slugify==8.0.1
pytorch-triton==2.0.0+b8b470bc59
pytz==2022.7.1
pyxdg==0.27
PyYAML==6.0
pyzmq==25.0.2
readchar==4.0.5
reportlab==3.6.8
requests==2.28.2
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986==1.5.0
rfc3986-validator==0.1.1
rich==13.3.2
rsa==4.9
scikit-learn==1.2.2
scipy==1.10.1
screen-resolution-extra==0.0.0
seaborn==0.12.2
SecretStorage==3.3.1
Send2Trash==1.8.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.4
ssh-import-id==5.11
stack-data==0.6.2
starlette==0.22.0
starsessions==1.3.0
swifter==1.3.4
sympy==1.11.1
systemd-python==234
tensorboard==2.12.0
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorboardX==2.6
terminado==0.17.1
text-unidecode==1.3
threadpoolctl==3.1.0
tinycss2==1.2.1
tomli==2.0.1
toolz==0.12.0
torch==2.0.0.dev20230301+cu117
torchaudio==2.0.0.dev20230302+cu117
torchmetrics==0.11.4
torchvision==0.15.0.dev20230302+cu117
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.5.0
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
ujson==5.7.0
unattended-upgrades==0.1
uri-template==1.2.0
urllib3==1.26.5
uvicorn==0.21.1
uvloop==0.17.0
wadllib==1.3.6
watchfiles==0.18.1
wcwidth==0.2.6
webcolors==1.12
webencodings==0.5.1
websocket-client==1.5.1
websockets==10.4
Werkzeug==2.2.3
widgetsnbextension==4.0.5
xdg==5
xkit==0.0.0
y-py==0.5.9
yarl==1.8.2
ypy-websocket==0.8.2
zipp==1.0.0
